+++ NEWS.md
@@ -7,6 +7,7 @@
 
 * Use annual programme data input instead of quarterly.
 * Add spectrum_region_code to area hierarchy.
+* Add advanced model run options
 * Add skeleton of model input validation
 
 # naomi 0.0.16