+++ NEWS.md
@@ -1,4 +1,3 @@
-
 # naomi 0.0.20
 
 * Support multiple PJNZ extracts in a zip folder.
@@ -13,6 +12,7 @@
 # naomi 0.0.18
 
 * Use odds ratios from Spectrum to model prevalence and ART coverage above age of data availability.
+* Save metadata alongside model outputs, when run via the hint run entrypoint (mrc-760)
 
 # naomi 0.0.17